Start unittesting flow.py
THIS IS WORK IN PROGRESS!

There are several issues with the Log decorator and unittesting.
While a decorated function is under test we need to access the
undecorated function because we want to assert the return value
of that function. That is done by copying the origin function to
an attribute in the Log Class, named _fn.
While testing the whole nacl-git.py functions, which are not
located inside a Class, that works out well.
We simply call the decorator objects ._fn attribute that contains
the origin function.

But: When we started testing the nacl-flow.py functions that resides
inside the Class NaclFlow we ran into problems.
The Log decorator Class works fine with the NaclFlow class regarding
decorating NaclFLow Class methods.
Methods are decorated.
That is done by implementing the descriptor protocol with the
__get__ Class method, that "overwrites" the "self" of Class Log with
self of Class NaclFlow.
We need that to ensure, that while decorating methods the origin
(decorated) attributes are still available.
But that leads to a significant problem:
There is no ._fn no more, the undecorated version of the origin
function is gone. OR: There is a possibility to do this we haven't
encountered yet.

As a workaround the decorator was implemented once more, almost
as a 1-1 copy but as a simple function, not as a Class. Inside
the function we copy the origin (say: decorated) function
to an attribute, that is accessible from outside.

Regarding unittesting NaclFlow that seems so solve the problem.
First implementations using 'log' instead of 'Log' are looking
good.

TODO: Further research is needed to decide which function to use
for all log decorations. There is no sense in using both.

So we have to write tests for NaclFlow to test the whole
Class (there are just 20% or so coverage) and then to
test, if we can move git.py from Log to log.
That requires the origin tests to adapt accordingly.