Add support for setting up the .nacl file
This is a rough yet not pretty implementiation of a setup routine
for the ~/.nacl file.
When a .nacl file is found its values are read and used as default
values. So `nacl-setup nacl` can be used to modify the .nacl file
as well.

The implementationhas its flaws because it is not flexible enough
to add new .nacl values easily.
Every new value has to be guided by a new block of code (and unittest).
This is a sign that the implementation is not sufficient.

But I was unable to came up with any easy-to-understand abstraction.
So it is not beautiful but understandable code I guess.