Make setup_nacl() more robust
In case that a .nacl is in place but has missing elements (like
'proxy') a KeyError was thrown.
Now there is a user_template with defaults which will act as
placeholders in case something is missing.
So we may provide some useful defaults here in case when we add
more configuration variables.